requirements.txt
django~=3.1.0
djangorestframework~=3.11.0
django-cors-headers
djangorestframework-jwt
django-filter